# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": " Api Documentation",
        "description": " Api Documentation",
        "contact": {
            "name": "Deni Akbar",
            "email": "dnxdenny@gmail.com"
        },
        "version": "1.0.0"
    },
    "paths": {
...
        "/api/images": {
            "get": {
                "tags": [
                    "Images"
                ],
                "summary": "Get list of images",
                "description": "Get list of images",
                "operationId": "getAllImages",
                "responses": {
                    "200": {
                        "description": "Success",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "status": {
                                            "type": "integer",
                                            "example": "200"
                                        },
                                        "data": {
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            },
            "post": {
                "tags": [
                    "Images"
                ],
                "summary": "Store Images ",
                "description": "Store Images ",
                "operationId": "storeImage",
                "requestBody": {
                    "required": true,
                    "content": {
                        "multipart/form-data": {
                            "schema": {
                                "required": [
                                    "file",
                                    "name",
                                    "enable"
                                ],
                                "properties": {
                                    "file": {
                                        "description": "file to upload",
                                        "type": "file"
                                    },
                                    "name": {
                                        "type": "string",
                                        "format": "string",
                                        "example": "File product"
                                    },
                                    "enable": {
                                        "type": "integer",
                                        "format": "integer",
                                        "example": "1"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Success",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "status": {
                                            "type": "integer",
                                            "example": ""
                                        },
                                        "data": {
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/images/{id}": {
            "get": {
                "tags": [
                    "Images"
                ],
                "summary": "Get Images Detail",
                "description": "Get Images Detail",
                "operationId": "showImages",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Id of Images",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Success",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "status_code": {
                                            "type": "integer",
                                            "example": "200"
                                        },
                                        "data": {
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            },
            "post": {
                "tags": [
                    "Images"
                ],
                "summary": "Update images",
                "description": "Update images",
                "operationId": "updateImage",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Id of Images",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "multipart/form-data": {
                            "schema": {
                                "required": [
                                    "file",
                                    "name",
                                    "enable"
                                ],
                                "properties": {
                                    "file": {
                                        "description": "file to upload",
                                        "type": "file"
                                    },
                                    "name": {
                                        "type": "string",
                                        "format": "string",
                                        "example": "File product"
                                    },
                                    "enable": {
                                        "type": "integer",
                                        "format": "integer",
                                        "example": "1"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Success",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "status_code": {
                                            "type": "integer",
                                            "example": "200"
                                        },
                                        "data": {
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            },
            "delete": {
                "tags": [
                    "Images"
                ],
                "summary": "Delete existing product",
                "description": "",
                "operationId": "deleteImage",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Image id",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Success",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "status_code": {
                                            "type": "integer",
                                            "example": "200"
                                        },
                                        "data": {
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
...
    }
}